backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Async support
httpx>=0.26.0
websockets>=12.0

# Utilities
python-dotenv>=1.0.0

# Parent package dependencies (shared with firefly_bill_sync)
requests>=2.28.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dateutil>=2.8.0
